Timer long/double error fix

+++ GoTrack/app/src/main/java/de/mobcom/group3/gotrack/Recording/Timer.java
@@ -87,7 +87,7 @@ public class Timer {
     }
 
     // return counted Time
-    public double getTime() {
+    public long getTime() {
         return time;
     }
